JS for all used functions
Ran in to an issue with vs code live server not recognizing the
external js script or the CSS, however when executed normally
there is no issue. So all JS is done in line of the index.html

So far I've solved how I want all of the functions to work, I just
need to create a way to start it all. From the documentation for
the project it seems like it's all done via console, however
adding an html button to push the start the game shouldn't be
too hard.

Actually, I believe the project will use the existing code and
then all the graphical elements will be added. At least creating
the function automation should be possible.

For any additional information, I've also created a scratchpaper.md
that contains some thoughts, issues, etc. While working through the
project.