Migrate script to js
I chose to rewrite all of the code
in the index.html into the myScripts.js
I also ended up changing all of the functions

Only thing left to do is to figure out the
getGameResults() function. The switch cases work
it's just that I'm not able to properly push
the window.confirm method through for a yes no
option for a new game and to display the results.