Completed Project
It took a lot of blood sweat and tears. Took a break
yesterday, but it's all finished.

All finished code is inside the myScripts.js,
the old original code is on the index.html, I don't recommend
looking at it.

If you want a look at some of the thought process and
issues I ran in to, a lot of it is on the scratchpaper.